chore: rename utils to Utils
