+++ src/Controls/GlobeControls.js
@@ -2,7 +2,7 @@
 import * as THREE from 'three';
 import AnimationPlayer, { Animation } from 'Core/AnimationPlayer';
 import Coordinates, { ellipsoidSizes } from 'Core/Geographic/Coordinates';
-import CameraUtils from 'utils/CameraUtils';
+import CameraUtils from 'Utils/CameraUtils';
 import StateControl from 'Controls/StateControl';
 
 // private members
