+++ src/Main.js
@@ -23,10 +23,10 @@ export { default as FirstPersonControls } from 'Controls/FirstPersonControls';
 export { default as PlanarControls } from 'Controls/PlanarControls';
 export { CONTROL_EVENTS } from 'Controls/GlobeControls';
 export { default as Feature2Mesh } from 'Converter/Feature2Mesh';
-export { default as FeaturesUtils } from 'utils/FeaturesUtils';
-export { default as DEMUtils } from 'utils/DEMUtils';
-export { default as CameraUtils } from 'utils/CameraUtils';
-export { default as OrientationUtils } from 'utils/OrientationUtils';
+export { default as FeaturesUtils } from 'Utils/FeaturesUtils';
+export { default as DEMUtils } from 'Utils/DEMUtils';
+export { default as CameraUtils } from 'Utils/CameraUtils';
+export { default as OrientationUtils } from 'Utils/OrientationUtils';
 export { default as ShaderChunk } from 'Renderer/Shader/ShaderChunk';
 export { getMaxColorSamplerUnitsCount } from 'Renderer/LayeredMaterial';
