+++ src/Utils/Utf8Decoder.js
